Split begrunnelse into smaller components
