+++ src/styled-components/alert.tsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components/macro';
+import { AlertStripeFeil } from 'nav-frontend-alertstriper';
+import AlertStripe from 'nav-frontend-alertstriper';
+
+export const KlageAlertStripeFeil = styled(AlertStripeFeil)`
+    && {
+        margin-bottom: 32px;
+    }
+`;
+
+export const KlageAlertStripe = styled(AlertStripe)`
+    && {
+        margin-bottom: 32px;
+    }
+`;
